@@ -1,15 +1,26 @@
 {
-  "name": "pdx-app",
+  "name": "sectioner-ui-library",
   "version": "0.1.0",
   "private": true,
+  "description": "Sectioner UI Library",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Chetan-KK/Sectioner-UI.git"
+  },
+  "author": "Chetan Khulage <chetankhulage.dev@gmail.com>",
+  "bugs": {
+    "url": "https://github.com/Chetan-KK/Sectioner-UI/issues"
+  },
+  "homepage": "https://github.com/Chetan-KK/Sectioner-UI#readme",
+  "license": "MIT",
   "dependencies": {
-    "next": "14.0.4",
+    "next": "14.2.2",
     "nextra": "^2.13.4",
     "nextra-theme-docs": "^2.13.4",
     "react": "^18",
@@ -19,11 +30,10 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "autoprefixer": "^10.0.1",
     "eslint": "^8",
-    "eslint-config-next": "14.0.4",
+    "eslint-config-next": "14.2.2",
     "postcss": "^8",
-    "tailwindcss": "^3.3.0",
+    "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
 }
