@@ -1,11 +1,10 @@
+import Image from "next/image";
 import Link from "next/link";
 
 export default function Home() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <Link href={"/docs/getting-started"} className="underline">
-        docs
-      </Link>
+      <Link href={'/docs/introduction'}>docs</Link>
     </main>
   );
 }
